Fix the issue of QD/DD lib; Fix the nan in binary tree orbital angle calculation; Add a few sample of triple systems

+++ sample/AR/ar.cxx
@@ -33,13 +33,13 @@ int main(int argc, char **argv){
     COMM::IOParams<int> print_precision(input_par_store, 14,   "print digital precision"); //print digital precision
     COMM::IOParams<int> nstep_max      (input_par_store, 1000000, "number of maximum (integrate/output) step for AR integration"); // maximum time step allown for tsyn integration
     COMM::IOParams<int> sym_order      (input_par_store, -6,   "Symplectic integrator order, should be even number"); // symplectic integrator order
-    COMM::IOParams<Float> energy_error (input_par_store, 1e-10,"relative energy error limit for AR"); // phase error requirement
-    COMM::IOParams<Float> time_error   (input_par_store, 0.0,  "time synchronization absolute error limit for AR","default is 0.25*dt-min"); // time synchronization error
-    COMM::IOParams<Float> time_zero    (input_par_store, 0.0,  "initial physical time");    // initial physical time
-    COMM::IOParams<Float> time_end     (input_par_store, 1.0,  "ending physical time"); // ending physical time
+    COMM::IOParams<double> energy_error (input_par_store, 1e-10,"relative energy error limit for AR"); // phase error requirement
+    COMM::IOParams<double> time_error   (input_par_store, 0.0,  "time synchronization absolute error limit for AR","default is 0.25*dt-min"); // time synchronization error
+    COMM::IOParams<double> time_zero    (input_par_store, 0.0,  "initial physical time");    // initial physical time
+    COMM::IOParams<double> time_end     (input_par_store, 1.0,  "ending physical time"); // ending physical time
     COMM::IOParams<int>   nstep        (input_par_store, 1000, "number of integration steps"); // total step size
-    COMM::IOParams<Float> s            (input_par_store, 0.0,  "step size, not physical time step","auto");    // step size
-    COMM::IOParams<Float> dt_min       (input_par_store, 1e-13,"minimum physical time step"); // minimum physical time step
+    COMM::IOParams<double> s            (input_par_store, 0.0,  "step size, not physical time step","auto");    // step size
+    COMM::IOParams<double> dt_min       (input_par_store, 1e-13,"minimum physical time step"); // minimum physical time step
     COMM::IOParams<int>   fix_step_option (input_par_store, -1, "always, later, none","auto"); // if true; use input fix step option
     COMM::IOParams<std::string> filename_par (input_par_store, "", "filename to load manager parameters","input name"); // par dumped filename
     bool load_flag=false; // if true; load dumped data
